Move benchmark to benchmark.py, reduce windows_example.py to minimum
Update docs with reference to example script
